Skip BBs without SSA representation in the Constant Propagation phase
In some cases the constant propagation optimization may get the MIR graph
where some of the BBs have no predecessors and do not transformed to the
SSA form. If such BB has operations on constants this may lead to segfault.
This patch adds the condition that will pass the only BBs with SSA.

Change-Id: I816d46b2492c5bd4748f983c3725b4798f9ebd68
Signed-off-by: Alexei Zavjalov <alexei.zavjalov@intel.com>

Conflicts:
	compiler/dex/mir_optimization.cc
	test/083-compiler-regressions/expected.txt
	test/083-compiler-regressions/src/Main.java